+++ src/lib/index.ts
@@ -6,3 +6,4 @@ export * from './paginator/index';
 export * from './spinner/index';
 export * from './dropdown/index';
 export * from './option/index';
+export * from './progress-bar/index';